Addition of font-awesome icons
An info icon added to show a user that more info is available on click.

A heart icon added to add favoriting functionality.

Heart icon styling done depending on the icon being clicked.

Functionality to display more data removed from the dog's image to the info icon.

Placeholder text added for properties without a value in the display-more-data container.